+++ cmake/onnxruntime_flatbuffers.cmake
@@ -1,0 +1,19 @@
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License.
+
+file(GLOB onnxruntime_flatbuffers_srcs CONFIGURE_DEPENDS
+    "${ONNXRUNTIME_ROOT}/core/flatbuffers/*.h"
+    "${ONNXRUNTIME_ROOT}/core/flatbuffers/*.cc"
+    )
+
+source_group(TREE ${REPO_ROOT} FILES ${onnxruntime_flatbuffers_srcs})
+
+add_library(onnxruntime_flatbuffers ${onnxruntime_flatbuffers_srcs})
+install(DIRECTORY ${PROJECT_SOURCE_DIR}/../include/onnxruntime/core/flatbuffers  DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/onnxruntime/core)
+onnxruntime_add_include_to_target(onnxruntime_flatbuffers onnx flatbuffers)
+if(onnxruntime_ENABLE_INSTRUMENT)
+  target_compile_definitions(onnxruntime_flatbuffers PUBLIC ONNXRUNTIME_ENABLE_INSTRUMENT)
+endif()
+target_include_directories(onnxruntime_flatbuffers PRIVATE ${ONNXRUNTIME_ROOT})
+add_dependencies(onnxruntime_flatbuffers ${onnxruntime_EXTERNAL_DEPENDENCIES})
+set_target_properties(onnxruntime_flatbuffers PROPERTIES FOLDER "ONNXRuntime")
