+++ cmake/onnxruntime_training.cmake
@@ -93,6 +93,7 @@ set(ONNXRUNTIME_LIBS
     onnxruntime_graph
     onnxruntime_common
     onnxruntime_mlas
+    onnxruntime_flatbuffers
 )
 
 if (onnxruntime_ENABLE_LANGUAGE_INTEROP_OPS)
