+++ onnxruntime/core/graph/graph_flatbuffers_utils.h
@@ -2,19 +2,40 @@
 // Licensed under the MIT License.
 
 #pragma once
-#include <core/graph/graph.h>
+
+namespace ONNX_NAMESPACE {
+class TensorProto;
+class AttributeProto;
+}  // namespace ONNX_NAMESPACE
+
+namespace flatbuffers {
+class FlatBufferBuilder;
+template <typename T>
+struct Offset;
+}  // namespace flatbuffers
 
 namespace onnxruntime {
+
+class Graph;
+class Node;
+
+namespace logging {
+class Logger;
+}
+
 namespace experimental {
+
+namespace fbs {
+struct Attribute;
+struct Tensor;
+}  // namespace fbs
+
 namespace utils {
 
-onnxruntime::common::Status SaveValueInfoOrtFormat(
-    flatbuffers::FlatBufferBuilder& builder, const ONNX_NAMESPACE::ValueInfoProto& value_info_proto,
-    flatbuffers::Offset<fbs::ValueInfo>& fbs_value_info) ORT_MUST_USE_RESULT;
-
+// TODO, add ORT_MUST_USE_RESULT when it is moved to a different header
 onnxruntime::common::Status SaveInitializerOrtFormat(
     flatbuffers::FlatBufferBuilder& builder, const ONNX_NAMESPACE::TensorProto& initializer,
-    flatbuffers::Offset<fbs::Tensor>& fbs_tensor) ORT_MUST_USE_RESULT;
+    flatbuffers::Offset<fbs::Tensor>& fbs_tensor);
 
 // Convert a given AttributeProto into fbs::Attribute
 // Note, we current do not support graphs, and sparse_tensor(s)
@@ -22,19 +43,12 @@ onnxruntime::common::Status SaveInitializerOrtFormat(
 //       instead of the GraphProto in attr_proto
 onnxruntime::common::Status SaveAttributeOrtFormat(
     flatbuffers::FlatBufferBuilder& builder, const ONNX_NAMESPACE::AttributeProto& attr_proto,
-    flatbuffers::Offset<fbs::Attribute>& fbs_attr, const onnxruntime::Graph* graph) ORT_MUST_USE_RESULT;
+    flatbuffers::Offset<fbs::Attribute>& fbs_attr, const onnxruntime::Graph* graph);
 
 #if defined(ENABLE_ORT_FORMAT_LOAD)
-inline void LoadStringFromOrtFormat(std::string& dst, const flatbuffers::String* fbs_string) {
-  if (fbs_string)
-    dst = fbs_string->c_str();
-}
 
 onnxruntime::common::Status LoadInitializerOrtFormat(
-    const fbs::Tensor& fbs_tensor, ONNX_NAMESPACE::TensorProto& initializer) ORT_MUST_USE_RESULT;
-
-onnxruntime::common::Status LoadValueInfoOrtFormat(
-    const fbs::ValueInfo& fbs_value_info, ONNX_NAMESPACE::ValueInfoProto& value_info_proto) ORT_MUST_USE_RESULT;
+    const fbs::Tensor& fbs_tensor, ONNX_NAMESPACE::TensorProto& initializer);
 
 // Load a give fbs::Attribute into AttributeProto
 // Note, If the attribute type is a graph, we will leave an empty graph in attr_proto,
@@ -43,7 +57,7 @@ onnxruntime::common::Status LoadAttributeOrtFormat(const fbs::Attribute& fbs_att
                                                    ONNX_NAMESPACE::AttributeProto& attr_proto,
                                                    std::unique_ptr<onnxruntime::Graph>& sub_graph,
                                                    Graph& graph, Node& node,
-                                                   const logging::Logger& logger) ORT_MUST_USE_RESULT;
+                                                   const logging::Logger& logger);
 
 #endif
