+++ onnxruntime/test/onnx/onnx_model_info.cc
@@ -6,17 +6,36 @@
 #include "re2/re2.h"
 #include "pb_helper.h"
 
-#if defined(ORT_MINIMAL_BUILD)
+#if defined(ENABLE_ORT_FORMAT_LOAD)
+
 #include <fstream>
-#include "core/graph/model.h"
-#include "core/common/logging/sinks/clog_sink.h"
+#include "core/flatbuffers/schema/ort.fbs.h"
+#include "core/flatbuffers/flatbuffers_utils.h"
 using namespace onnxruntime::experimental;
-using namespace onnxruntime::logging;
+
 #endif
 
 using namespace onnxruntime;
 
+OnnxModelInfo::OnnxModelInfo(_In_ const PATH_CHAR_TYPE* model_url, bool is_ort_model)
+    : model_url_(model_url) {
+  if (is_ort_model) {
+#if defined(ENABLE_ORT_FORMAT_LOAD)
+    InitOrtModelInfo(model_url);
+#else
+    ORT_THROW("ort model is not supported in this build");
+#endif
+  } else {
 #if !defined(ORT_MINIMAL_BUILD)
+    InitOnnxModelInfo(model_url);
+#else
+    ORT_THROW("onnx model is not supported in this build");
+#endif
+  }
+}
+
+#if !defined(ORT_MINIMAL_BUILD)
+
 static constexpr int protobuf_block_size_in_bytes = 4 * 1024 * 1024;
 template <typename T>
 static void RepeatedPtrFieldToVector(const ::google::protobuf::RepeatedPtrField<T>& input_value_info,
@@ -26,9 +45,7 @@ static void RepeatedPtrFieldToVector(const ::google::protobuf::RepeatedPtrField<
   }
 }
 
-OnnxModelInfo::OnnxModelInfo(_In_ const PATH_CHAR_TYPE* model_url)
-    : BaseModelInfo(model_url) {
-  // parse model
+void OnnxModelInfo::InitOnnxModelInfo(_In_ const PATH_CHAR_TYPE* model_url) {  // parse model
   int model_fd;
   auto st = Env::Default().FileOpenRd(model_url, model_fd);
   if (!st.IsOK()) {
@@ -81,10 +98,11 @@ OnnxModelInfo::OnnxModelInfo(_In_ const PATH_CHAR_TYPE* model_url)
   RepeatedPtrFieldToVector(graph.output(), output_value_info_);
 }
 
-#else
+#endif  // #if !defined(ORT_MINIMAL_BUILD)
 
-OrtModelInfo::OrtModelInfo(_In_ const PATH_CHAR_TYPE* model_url)
-    : BaseModelInfo(model_url) {
+#if defined(ENABLE_ORT_FORMAT_LOAD)
+
+void OnnxModelInfo::InitOrtModelInfo(_In_ const PATH_CHAR_TYPE* model_url) {
   std::vector<uint8_t> bytes;
   size_t num_bytes = 0;
   const auto model_location = ToWideString(model_url);
@@ -92,6 +110,9 @@ OrtModelInfo::OrtModelInfo(_In_ const PATH_CHAR_TYPE* model_url)
   bytes.resize(num_bytes);
   std::ifstream bytes_stream(model_location, std::ifstream::in | std::ifstream::binary);
   bytes_stream.read(reinterpret_cast<char*>(bytes.data()), num_bytes);
+
+  // TODO use ort format version here?
+  onnx_commit_tag_ = TestModelInfo::unknown_version;
 
   // TODO, verify it is a valid ort format
   // TODO, version matches the ORT version
@@ -103,34 +124,56 @@ OrtModelInfo::OrtModelInfo(_In_ const PATH_CHAR_TYPE* model_url)
   if (nullptr == fbs_model)
     ORT_THROW("Missing Model. Invalid ORT format model.");
 
-  std::unique_ptr<Model> model;
-  // create scoped manager so sink gets destroyed once done
-  {
-    LoggingManager manager{std::unique_ptr<ISink>{new CLogSink{}}, Severity::kVERBOSE, false,
-                           LoggingManager::InstanceType::Temporal};
+  const auto* fbs_graph = fbs_model->graph();
+  if (nullptr == fbs_graph)
+    ORT_THROW("Missing Graph. Invalid ORT format model.");
 
-    auto logger = manager.CreateLogger("CreateOrtModelInfo");
-    ORT_THROW_IF_ERROR(Model::LoadFromOrtFormat(*fbs_model, *logger, model));
-  }
-
-  // TODO use ort format version here?
-  onnx_commit_tag_ = TestModelInfo::unknown_version;
-
-  Graph& graph = model->MainGraph();
-  for (const auto entry : graph.DomainToVersionMap()) {
+  std::unordered_map<std::string, int> _opset_import;
+  ORT_THROW_IF_ERROR(experimental::utils::LoadOpsetImportOrtFormat(fbs_model->opset_import(), _opset_import));
+  for (const auto& entry : _opset_import)
     domain_to_version_[entry.first] = entry.second;
+
+  // Load all node args from fbs_graph
+  std::unordered_map<std::string, ONNX_NAMESPACE::ValueInfoProto> _node_args;
+  auto fbs_node_args = fbs_graph->node_args();
+  if (fbs_node_args) {
+    _node_args.reserve(fbs_node_args->size());
+    for (const auto* fbs_value_info : *fbs_node_args) {
+      if (nullptr == fbs_value_info)
+        ORT_THROW("NodeArg is missing. Invalid ORT format model.");
+      ONNX_NAMESPACE::ValueInfoProto node_arg_info;
+      ORT_THROW_IF_ERROR(experimental::utils::LoadValueInfoOrtFormat(*fbs_value_info, node_arg_info));
+      // NodeArg ctor is private, cannot use make_unique
+      _node_args[fbs_value_info->name()->str()] = std::move(node_arg_info);
+    }
   }
 
-  if (graph.NumberOfNodes() == 1) {
-    node_name_ = graph.Nodes().cbegin()->OpType();
-  }
+  if (fbs_graph->nodes() && fbs_graph->nodes()->size() == 1) {
+    const auto* node = fbs_graph->nodes()->Get(0);
+    if (!node)
+      ORT_THROW("Missing Node. Invalid ORT format model.");
 
-  for (const auto* node_arg : graph.GetInputs()) {
-    input_value_info_.push_back(node_arg->ToProto());
-  }
+    if (!node->op_type())
+      ORT_THROW("Missing op_type. Invalid ORT format model.");
 
-  for (const auto* node_arg : graph.GetOutputs()) {
-    output_value_info_.push_back(node_arg->ToProto());
+    node_name_ = node->op_type()->str();
   }
+  // Load input and output node args
+  auto add_node_args = [&](const flatbuffers::Vector<flatbuffers::Offset<flatbuffers::String>>* fbs_node_args,
+                           std::vector<ONNX_NAMESPACE::ValueInfoProto> node_args) -> Status {
+    if (fbs_node_args != nullptr) {
+      node_args.reserve(fbs_node_args->size());
+      for (const auto* fbs_node_arg_name : *fbs_node_args) {
+        if (!fbs_node_arg_name)
+          ORT_THROW("NodeArg Name is missing. Invalid ORT format model.");
+        node_args.push_back(_node_args.at(fbs_node_arg_name->str()));
+      }
+    }
+    return Status::OK();
+  };
+
+  ORT_THROW_IF_ERROR(add_node_args(fbs_graph->inputs(), input_value_info_));
+  ORT_THROW_IF_ERROR(add_node_args(fbs_graph->outputs(), output_value_info_));
 }
-#endif  //#if !defined(ORT_MINIMAL_BUILD)
+
+#endif  //#if defined(ENABLE_ORT_FORMAT_LOAD)
